+++ src/screens/register-admin/Form.tsx
@@ -8,7 +8,7 @@ import {
   FormMessage
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { registerAdminService } from "@/services/accounts/accounts.services";
+import { registerAdminService } from "@/services/accounts/register-admin.service";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
